chore: rename posts page
