+++ src/app/posts/page.tsx
@@ -1,4 +1,4 @@
-import { PostsPage } from "@/pages/posts/ui/page";
+import { PostsPage } from "@/pages/posts/ui/PostsPage";
 
 export default function Posts() {
   return <PostsPage />;
