+++ src/pages/posts/ui/PostsPage.tsx
